[CHORE] #53 : editProfile -> profileAllergy rename
