+++ client/app/src/main/res/layout/fragment_profile_allergy.xml
@@ -4,7 +4,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.example.healthc.presentation.profile.edit_profile.EditProfileViewModel"
+            type="com.example.healthc.presentation.profile.profile_allergy.ProfileAllergyViewModel"
             />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
